Switch pool.apply to direct call_command calls; add TODOs
Latest version of Django is causing some bizarre errors and isn't recognizing the commands when call_command is used via multiprocessing. Need to fix this; multiprocessing is currently not functioning. For now, we'll switch away from pool.apply and call call_command directly so the functions at least work when num_cores=1